Refactor: 폴더별로 타입 분리 (관심사의 분리)
