+++ src/components/Card/Card.types.ts
@@ -1,0 +1,12 @@
+export interface CardProps {
+  onSale: boolean;
+  workImageUrl: string;
+  workTitle: string;
+  likeCount: number;
+  viewCount: number;
+  commentCount: number;
+  profileImageUrl: string;
+  authorName: string;
+  workUrl: string;
+  authorUrl: string;
+}
