+++ src/components/SideBar/SideBar.tsx
@@ -1,20 +1,12 @@
 import '@/styles/tailwind.css';
-import Image, { StaticImageData } from 'next/image';
+import Image from 'next/image';
 import Link from 'next/link';
 import EditIcon from './EditIcon';
 import LinkIcon from './LinkIcon';
 import ProfileIcon from './ProfileIcon';
+import { SideBarProps } from './SideBar.types';
 
-interface Props {
-  name: string;
-  role: string;
-  description: string;
-  likes: number;
-  followers: number;
-  image?: string | StaticImageData;
-}
-
-function SideBar({ name, role, description, likes, followers, image }: Props) {
+function SideBar({ name, role, description, likes, followers, image }: SideBarProps) {
   return (
     <div className="fixed left-36 top-110 h-648 w-260 rounded-sm">
       <div className="absolute -top-10 left-1/2 z-10 h-120 w-120 -translate-x-1/2 transform rounded-full">
