+++ src/components/SideBar/SideBar.types.ts
@@ -1,0 +1,10 @@
+import { StaticImageData } from 'next/image';
+
+export interface SideBarProps {
+  name: string;
+  role: string;
+  description: string;
+  likes: number;
+  followers: number;
+  image?: string | StaticImageData;
+}
